Randomize Computer Selection and comparing values
Implemented the computers choice into the code, and now the beginning
process of comparing rock vs paper vs scissors. Because strings in an
array hold the value of a specific number, these numbers are going to be
compared. Same values like 1-1 are already omitted (draw); so now just
work with 0-2, 0-1.