Compare values, select winner
Whatever the user types and computer randomly selects, the two are
converted into a number--the position where they're in . in the array--
and converted back into a string after being added together. This value
is then compared to compWin (array) to see whether or not the said value
is included. If yes, the computer wins. If not, it checks myWin and if
that array includes the value, then user wins. (If neither, error)