Finish game, scoring system
The majority of the code, from choosing an answer to comparing the
values, was put under a function, round(). After each round, either
the User or the Computer wins a point (incrementing). Whoever wins 2
points, wins the game and the code ends. (Only one problem, the winner
is displayed twice.)