Tie Breaker, Fix duplicate "<> wins" message
Finished (for now). Implemented a Tie Breaker, so if both players
have their score equal 2, then a 3rd and final round will begin with the
winner being the overall winner.

Also fixed the issue where, if you had a draw, the winner would be
displayed x number_of_draws. Because everything is under the function
round(), by invoking it, everything is run through again rather than
a specific portion. And because there were two instances where round()
was invoked, that I believe is where the problem occured. (I removed
the round() under the if statement that states there was a draw, and
it removed the problem).